timeSeries alignment corrected for avx, multiplyadd added, performance tests done for odd input, enabled timeseries tests
