+++ src/Microsoft.ML.CpuMath/CpuMathUtils.netstandard.cs
@@ -5,6 +5,7 @@
 using Microsoft.ML.Runtime.Internal.CpuMath.Core;
 using System;
 using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
 
 namespace Microsoft.ML.Runtime.Internal.CpuMath
 {
@@ -12,63 +13,417 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
     internal static partial class CpuMathUtils
     {
         // The count of bytes in Vector128<T>, corresponding to _cbAlign in AlignedArray
-        private const int Vector128Alignment = 16;
+        private const int CbAlign = 16;
 
         [MethodImplAttribute(MethodImplOptions.AggressiveInlining)]
         public static int GetVectorAlignment()
-            => Vector128Alignment;
+            => CbAlign;
 
-        public static void MatrixTimesSource(bool transpose, AlignedArray matrix, AlignedArray source, AlignedArray destination, int stride) => SseUtils.MatTimesSrc(transpose, matrix, source, destination, stride);
+        private static bool Compat(AlignedArray a)
+        {
+            Contracts.AssertValue(a);
+            Contracts.Assert(a.Size > 0);
+            return a.CbAlign == CbAlign;
+        }
 
-        public static void MatrixTimesSource(AlignedArray matrix, ReadOnlySpan<int> rgposSrc, AlignedArray sourceValues,
-            int posMin, int iposMin, int iposLimit, AlignedArray destination, int stride) => SseUtils.MatTimesSrc(matrix, rgposSrc, sourceValues, posMin, iposMin, iposLimit, destination, stride);
+        private static unsafe float* Ptr(AlignedArray a, float* p)
+        {
+            Contracts.AssertValue(a);
+            float* q = p + a.GetBase((long)p);
+            Contracts.Assert(((long)q & (CbAlign - 1)) == 0);
+            return q;
+        }
 
-        public static void Add(float value, Span<float> destination) => SseUtils.Add(value, destination);
+        public static void MatrixTimesSource(bool tran, AlignedArray mat, AlignedArray src, AlignedArray dst, int crun)
+        {
+            Contracts.Assert(Compat(mat));
+            Contracts.Assert(Compat(src));
+            Contracts.Assert(Compat(dst));
+            Contracts.Assert(mat.Size == dst.Size * src.Size);
 
-        public static void Scale(float value, Span<float> destination) => SseUtils.Scale(value, destination);
+            unsafe
+            {
+                fixed (float* pmat = &mat.Items[0])
+                fixed (float* psrc = &src.Items[0])
+                fixed (float* pdst = &dst.Items[0])
+                {
+                    if (!tran)
+                    {
+                        Contracts.Assert(0 <= crun && crun <= dst.Size);
+                        Thunk.MatMul(Ptr(mat, pmat), Ptr(src, psrc), Ptr(dst, pdst), crun, src.Size);
+                    }
+                    else
+                    {
+                        Contracts.Assert(0 <= crun && crun <= src.Size);
+                        Thunk.MatMulTran(Ptr(mat, pmat), Ptr(src, psrc), Ptr(dst, pdst), dst.Size, crun);
+                    }
+                }
+            }
+        }
 
-        public static void Scale(float value, ReadOnlySpan<float> source, Span<float> destination, int count) => SseUtils.Scale(value, source, destination, count);
+        public static void MatrixTimesSource(AlignedArray mat, ReadOnlySpan<int> rgposSrc, AlignedArray srcValues,
+            int posMin, int iposMin, int iposLim, AlignedArray dst, int crun)
+        {
+            Contracts.Assert(Compat(mat));
+            Contracts.Assert(Compat(srcValues));
+            Contracts.Assert(Compat(dst));
+            Contracts.Assert(0 <= iposMin && iposMin <= iposLim && iposLim <= rgposSrc.Length);
+            Contracts.Assert(mat.Size == dst.Size * srcValues.Size);
 
-        public static void ScaleAdd(float value, float addend, Span<float> destination) => SseUtils.ScaleAdd(value, addend, destination);
+            if (iposMin >= iposLim)
+            {
+                dst.ZeroItems();
+                return;
+            }
+            Contracts.AssertNonEmpty(rgposSrc);
+            unsafe
+            {
+                fixed (float* pdst = &dst.Items[0])
+                fixed (float* pmat = &mat.Items[0])
+                fixed (float* psrc = &srcValues.Items[0])
+                fixed (int* ppossrc = &rgposSrc[0])
+                {
+                    Contracts.Assert(0 <= crun && crun <= dst.Size);
+                    Thunk.MatMulP(Ptr(mat, pmat), ppossrc, Ptr(srcValues, psrc), posMin, iposMin, iposLim, Ptr(dst, pdst), crun, srcValues.Size);
+                }
+            }
+        }
 
-        public static void AddScale(float value, ReadOnlySpan<float> source, Span<float> destination, int count) => SseUtils.AddScale(value, source, destination, count);
+        // dst += a
+        public static void Add(float a, Span<float> dst)
+        {
+            Contracts.AssertNonEmpty(dst);
 
-        public static void AddScale(float value, ReadOnlySpan<float> source, ReadOnlySpan<int> indices, Span<float> destination, int count) => SseUtils.AddScale(value, source, indices, destination, count);
+            unsafe
+            {
+                fixed (float* pdst = &MemoryMarshal.GetReference(dst))
+                    Thunk.AddScalarU(a, pdst, dst.Length);
+            }
+        }
 
-        public static void AddScaleCopy(float value, ReadOnlySpan<float> source, ReadOnlySpan<float> destination, Span<float> res, int count) => SseUtils.AddScaleCopy(value, source, destination, res, count);
+        public static void Scale(float a, Span<float> dst)
+        {
+            Contracts.AssertNonEmpty(dst);
 
-        public static void Add(ReadOnlySpan<float> source, Span<float> destination, int count) => SseUtils.Add(source, destination, count);
+            unsafe
+            {
+                fixed (float* pd = &MemoryMarshal.GetReference(dst))
+                    Thunk.Scale(a, pd, dst.Length);
+            }
+        }
 
-        public static void Add(ReadOnlySpan<float> source, ReadOnlySpan<int> indices, Span<float> destination, int count) => SseUtils.Add(source, indices, destination, count);
+        // dst = a * src
+        public static void Scale(float a, ReadOnlySpan<float> src, Span<float> dst, int count)
+        {
+            Contracts.AssertNonEmpty(src);
+            Contracts.Assert(0 < count && count <= src.Length);
+            Contracts.AssertNonEmpty(dst);
+            Contracts.Assert(count <= dst.Length);
 
-        public static void MulElementWise(ReadOnlySpan<float> left, ReadOnlySpan<float> right, Span<float> destination, int count) => SseUtils.MulElementWise(left, right, destination, count);
+            unsafe
+            {
+                fixed (float* psrc = &MemoryMarshal.GetReference(src))
+                fixed (float* pdst = &MemoryMarshal.GetReference(dst))
+                {
+                    Thunk.ScaleSrcU(a, psrc, pdst, count);
+                }
+            }
+        }
 
-        public static float Sum(ReadOnlySpan<float> source) => SseUtils.Sum(source);
+        // dst[i] = a * (dst[i] + b)
+        public static void ScaleAdd(float a, float b, Span<float> dst)
+        {
+            Contracts.AssertNonEmpty(dst);
 
-        public static float SumSq(ReadOnlySpan<float> source) => SseUtils.SumSq(source);
+            unsafe
+            {
+                fixed (float* pdst = &MemoryMarshal.GetReference(dst))
+                    Thunk.ScaleAddU(a, b, pdst, dst.Length);
+            }
+        }
 
-        public static float SumSq(float mean, ReadOnlySpan<float> source) => SseUtils.SumSq(mean, source);
+        public static void AddScale(float a, ReadOnlySpan<float> src, Span<float> dst, int count)
+        {
+            Contracts.AssertNonEmpty(src);
+            Contracts.Assert(0 < count && count <= src.Length);
+            Contracts.AssertNonEmpty(dst);
+            Contracts.Assert(count <= dst.Length);
 
-        public static float SumAbs(ReadOnlySpan<float> source) => SseUtils.SumAbs(source);
+            unsafe
+            {
+                fixed (float* psrc = &MemoryMarshal.GetReference(src))
+                fixed (float* pdst = &MemoryMarshal.GetReference(dst))
+                    Thunk.AddScaleU(a, psrc, pdst, count);
+            }
+        }
 
-        public static float SumAbs(float mean, ReadOnlySpan<float> source) => SseUtils.SumAbs(mean, source);
+        public static void AddScale(float a, ReadOnlySpan<float> src, ReadOnlySpan<int> indices, Span<float> dst, int count)
+        {
+            Contracts.AssertNonEmpty(src);
+            Contracts.Assert(0 < count && count <= src.Length);
+            Contracts.AssertNonEmpty(indices);
+            Contracts.Assert(count <= indices.Length);
+            Contracts.AssertNonEmpty(dst);
+            Contracts.Assert(count < dst.Length);
 
-        public static float MaxAbs(ReadOnlySpan<float> source) => SseUtils.MaxAbs(source);
+            unsafe
+            {
+                fixed (float* psrc = &MemoryMarshal.GetReference(src))
+                fixed (int* pi = &MemoryMarshal.GetReference(indices))
+                fixed (float* pdst = &MemoryMarshal.GetReference(dst))
+                    Thunk.AddScaleSU(a, psrc, pi, pdst, count);
+            }
+        }
 
-        public static float MaxAbsDiff(float mean, ReadOnlySpan<float> source) => SseUtils.MaxAbsDiff(mean, source);
+        public static void AddScaleCopy(float a, ReadOnlySpan<float> src, ReadOnlySpan<float> dst, Span<float> res, int count)
+        {
+            Contracts.AssertNonEmpty(dst);
+            Contracts.Assert(0 < count && count <= dst.Length);
+            Contracts.AssertNonEmpty(src);
+            Contracts.Assert(count <= src.Length);
+            Contracts.AssertNonEmpty(res);
+            Contracts.Assert(count <= res.Length);
 
-        public static float DotProductDense(ReadOnlySpan<float> left, ReadOnlySpan<float> right, int count) => SseUtils.DotProductDense(left, right, count);
+            unsafe
+            {
+                fixed (float* pdst = &MemoryMarshal.GetReference(dst))
+                fixed (float* psrc = &MemoryMarshal.GetReference(src))
+                fixed (float* pres = &MemoryMarshal.GetReference(res))
+                    Thunk.AddScaleCopyU(a, psrc, pdst, pres, count);
+            }
+        }
 
-        public static float DotProductSparse(ReadOnlySpan<float> left, ReadOnlySpan<float> right, ReadOnlySpan<int> indices, int count) => SseUtils.DotProductSparse(left, right, indices, count);
+        public static void Add(ReadOnlySpan<float> src, Span<float> dst, int count)
+        {
+            Contracts.AssertNonEmpty(src);
+            Contracts.Assert(0 < count && count <= src.Length);
+            Contracts.AssertNonEmpty(dst);
+            Contracts.Assert(count <= dst.Length);
 
-        public static float L2DistSquared(ReadOnlySpan<float> left, ReadOnlySpan<float> right, int count) => SseUtils.L2DistSquared(left, right, count);
+            unsafe
+            {
+                fixed (float* ps = &MemoryMarshal.GetReference(src))
+                fixed (float* pd = &MemoryMarshal.GetReference(dst))
+                    Thunk.AddU(ps, pd, count);
+            }
+        }
 
-        public static void ZeroMatrixItems(AlignedArray destination, int ccol, int cfltRow, int[] indices) => SseUtils.ZeroMatrixItems(destination, ccol, cfltRow, indices);
+        public static void Add(ReadOnlySpan<float> src, ReadOnlySpan<int> indices, Span<float> dst, int count)
+        {
+            Contracts.AssertNonEmpty(src);
+            Contracts.Assert(0 < count && count <= src.Length);
+            Contracts.AssertNonEmpty(indices);
+            Contracts.Assert(count <= indices.Length);
+            Contracts.AssertNonEmpty(dst);
+            Contracts.Assert(count < dst.Length);
 
-        public static void SdcaL1UpdateDense(float primalUpdate, int count, ReadOnlySpan<float> source, float threshold, Span<float> v, Span<float> w)
-            => SseUtils.SdcaL1UpdateDense(primalUpdate, count, source, threshold, v, w);
+            unsafe
+            {
+                fixed (float* ps = &MemoryMarshal.GetReference(src))
+                fixed (int* pi = &MemoryMarshal.GetReference(indices))
+                fixed (float* pd = &MemoryMarshal.GetReference(dst))
+                    Thunk.AddSU(ps, pi, pd, count);
+            }
+        }
+
+        public static void MulElementWise(ReadOnlySpan<float> src1, ReadOnlySpan<float> src2, Span<float> dst, int count)
+        {
+            Contracts.AssertNonEmpty(src1);
+            Contracts.Assert(0 < count && count <= src1.Length);
+            Contracts.AssertNonEmpty(src2);
+            Contracts.Assert(0 < count && count <= src2.Length);
+            Contracts.AssertNonEmpty(dst);
+            unsafe
+            {
+                fixed (float* ps1 = &MemoryMarshal.GetReference(src1))
+                fixed (float* ps2 = &MemoryMarshal.GetReference(src2))
+                fixed (float* pd = &MemoryMarshal.GetReference(dst))
+                    Thunk.MulElementWiseU(ps1, ps2, pd, count);
+            }
+        }
+
+        public static float Sum(ReadOnlySpan<float> src)
+        {
+            Contracts.AssertNonEmpty(src);
+
+            unsafe
+            {
+                fixed (float* psrc = &MemoryMarshal.GetReference(src))
+                    return Thunk.Sum(psrc, src.Length);
+            }
+        }
+
+        public static float SumSq(ReadOnlySpan<float> src)
+        {
+            Contracts.AssertNonEmpty(src);
+
+            unsafe
+            {
+                fixed (float* psrc = &MemoryMarshal.GetReference(src))
+                    return Thunk.SumSqU(psrc, src.Length);
+            }
+        }
+
+        public static float SumSq(float mean, ReadOnlySpan<float> src)
+        {
+            Contracts.AssertNonEmpty(src);
+
+            unsafe
+            {
+                fixed (float* psrc = &MemoryMarshal.GetReference(src))
+                    return (mean == 0 ? Thunk.SumSqU(psrc, src.Length) : Thunk.SumSqDiffU(mean, psrc, src.Length));
+            }
+        }
+
+        public static float SumAbs(ReadOnlySpan<float> src)
+        {
+            Contracts.AssertNonEmpty(src);
+
+            unsafe
+            {
+                fixed (float* psrc = &MemoryMarshal.GetReference(src))
+                    return Thunk.SumAbsU(psrc, src.Length);
+            }
+        }
+
+        public static float SumAbs(float mean, ReadOnlySpan<float> src)
+        {
+            Contracts.AssertNonEmpty(src);
+
+            unsafe
+            {
+                fixed (float* psrc = &MemoryMarshal.GetReference(src))
+                    return (mean == 0 ? Thunk.SumAbsU(psrc, src.Length) : Thunk.SumAbsDiffU(mean, psrc, src.Length));
+            }
+        }
+
+        public static float MaxAbs(ReadOnlySpan<float> src)
+        {
+            Contracts.AssertNonEmpty(src);
+
+            unsafe
+            {
+                fixed (float* psrc = &MemoryMarshal.GetReference(src))
+                    return Thunk.MaxAbsU(psrc, src.Length);
+            }
+        }
+
+        public static float MaxAbsDiff(float mean, ReadOnlySpan<float> src)
+        {
+            Contracts.AssertNonEmpty(src);
+
+            unsafe
+            {
+                fixed (float* psrc = &MemoryMarshal.GetReference(src))
+                    return Thunk.MaxAbsDiffU(mean, psrc, src.Length);
+            }
+        }
+
+        public static float DotProductDense(ReadOnlySpan<float> a, ReadOnlySpan<float> b, int count)
+        {
+            Contracts.AssertNonEmpty(a);
+            Contracts.AssertNonEmpty(b);
+            Contracts.Assert(0 < count);
+            Contracts.Assert(a.Length >= count);
+            Contracts.Assert(b.Length >= count);
+
+            unsafe
+            {
+                fixed (float* pa = &MemoryMarshal.GetReference(a))
+                fixed (float* pb = &MemoryMarshal.GetReference(b))
+                    return Thunk.DotU(pa, pb, count);
+            }
+        }
+
+        public static float DotProductSparse(ReadOnlySpan<float> a, ReadOnlySpan<float> b, ReadOnlySpan<int> indices, int count)
+        {
+            Contracts.AssertNonEmpty(a);
+            Contracts.AssertNonEmpty(b);
+            Contracts.Assert(0 < count);
+            Contracts.Assert(count < a.Length);
+            Contracts.Assert(count <= b.Length);
+            Contracts.Assert(count <= indices.Length);
+
+            unsafe
+            {
+                fixed (float* pa = &MemoryMarshal.GetReference(a))
+                fixed (float* pb = &MemoryMarshal.GetReference(b))
+                fixed (int* pi = &MemoryMarshal.GetReference(indices))
+                    return Thunk.DotSU(pa, pb, pi, count);
+            }
+        }
+
+        public static float L2DistSquared(ReadOnlySpan<float> a, ReadOnlySpan<float> b, int count)
+        {
+            Contracts.AssertNonEmpty(a);
+            Contracts.AssertNonEmpty(b);
+            Contracts.Assert(0 < count && count <= a.Length);
+            Contracts.Assert(count <= b.Length);
+
+            unsafe
+            {
+                fixed (float* pa = &MemoryMarshal.GetReference(a))
+                fixed (float* pb = &MemoryMarshal.GetReference(b))
+                    return Thunk.Dist2(pa, pb, count);
+            }
+        }
+
+        public static void ZeroMatrixItems(AlignedArray dst, int ccol, int cfltRow, int[] indices)
+        {
+            Contracts.Assert(0 < ccol && ccol <= cfltRow);
+
+            unsafe
+            {
+                fixed (float* pdst = &dst.Items[0])
+                fixed (int* pi = &indices[0])
+                {
+                    if (ccol == cfltRow)
+                        Thunk.ZeroItemsU(Ptr(dst, pdst), dst.Size, pi, indices.Length);
+                    else
+                        Thunk.ZeroMatrixItemsCore(Ptr(dst, pdst), dst.Size, ccol, cfltRow, pi, indices.Length);
+                }
+            }
+        }
+
+        public static void SdcaL1UpdateDense(float primalUpdate, int count, ReadOnlySpan<float> src, float threshold, Span<float> v, Span<float> w)
+        {
+            Contracts.AssertNonEmpty(src);
+            Contracts.Assert(count <= src.Length);
+            Contracts.AssertNonEmpty(v);
+            Contracts.Assert(count <= v.Length);
+            Contracts.AssertNonEmpty(w);
+            Contracts.Assert(count <= w.Length);
+            Contracts.Assert(count > 0);
+
+            unsafe
+            {
+                fixed (float* psrc = &MemoryMarshal.GetReference(src))
+                fixed (float* pd1 = &MemoryMarshal.GetReference(v))
+                fixed (float* pd2 = &MemoryMarshal.GetReference(w))
+                    Thunk.SdcaL1UpdateU(primalUpdate, psrc, threshold, pd1, pd2, count);
+            }
+        }
 
         public static void SdcaL1UpdateSparse(float primalUpdate, int count, ReadOnlySpan<float> source, ReadOnlySpan<int> indices, float threshold, Span<float> v, Span<float> w)
-            => SseUtils.SdcaL1UpdateSparse(primalUpdate, count, source, indices, threshold, v, w);
+        {
+            Contracts.AssertNonEmpty(source);
+            Contracts.Assert(count <= source.Length);
+            Contracts.AssertNonEmpty(indices);
+            Contracts.Assert(count <= indices.Length);
+            Contracts.AssertNonEmpty(v);
+            Contracts.Assert(count <= v.Length);
+            Contracts.AssertNonEmpty(w);
+            Contracts.Assert(count <= w.Length);
+            Contracts.Assert(count > 0);
+
+            unsafe
+            {
+                fixed (float* psrc = &MemoryMarshal.GetReference(source))
+                fixed (int* pi = &MemoryMarshal.GetReference(indices))
+                fixed (float* pd1 = &MemoryMarshal.GetReference(v))
+                fixed (float* pd2 = &MemoryMarshal.GetReference(w))
+                    Thunk.SdcaL1UpdateSU(primalUpdate, psrc, pi, threshold, pd1, pd2, count);
+            }
+        }
     }
 }
