+++ test/Microsoft.ML.CpuMath.PerformanceTests/NativePerformanceTests.cs
@@ -6,17 +6,28 @@ using System;
 using BenchmarkDotNet.Attributes;
 using BenchmarkDotNet.Running;
 using Microsoft.ML.Runtime.Internal.CpuMath;
+using Microsoft.ML.Runtime.Internal.CpuMath.Core;
 
 namespace Microsoft.ML.CpuMath.PerformanceTests
 {
     public class NativePerformanceTests : PerformanceTests
     {
+        private const int CbAlign = 16;
+
+        private static unsafe float* Ptr(AlignedArray a, float* p)
+        {
+            Contracts.AssertValue(a);
+            float* q = p + a.GetBase((long)p);
+            Contracts.Assert(((long)q & (CbAlign - 1)) == 0);
+            return q;
+        }
+
         [Benchmark]
         public unsafe void AddScalarU()
         {
             fixed (float* pdst = dst)
             {
-                CpuMathNativeUtils.AddScalarU(DefaultScale, pdst, Length);
+                Thunk.AddScalarU(DefaultScale, pdst, Length);
             }
         }
 
@@ -25,7 +36,7 @@ namespace Microsoft.ML.CpuMath.PerformanceTests
         {
             fixed (float* pdst = dst)
             {
-                CpuMathNativeUtils.Scale(DefaultScale, pdst, Length);
+                Thunk.Scale(DefaultScale, pdst, Length);
             }
         }
 
@@ -35,7 +46,7 @@ namespace Microsoft.ML.CpuMath.PerformanceTests
             fixed (float* psrc = src)
             fixed (float* pdst = dst)
             {
-                CpuMathNativeUtils.ScaleSrcU(DefaultScale, psrc, pdst, Length);
+                Thunk.ScaleSrcU(DefaultScale, psrc, pdst, Length);
             }
         }
 
@@ -44,7 +55,7 @@ namespace Microsoft.ML.CpuMath.PerformanceTests
         {
             fixed (float* pdst = dst)
             {
-                CpuMathNativeUtils.ScaleAddU(DefaultScale, DefaultScale, pdst, Length);
+                Thunk.ScaleAddU(DefaultScale, DefaultScale, pdst, Length);
             }
         }
 
@@ -54,7 +65,7 @@ namespace Microsoft.ML.CpuMath.PerformanceTests
             fixed (float* psrc = src)
             fixed (float* pdst = dst)
             {
-                CpuMathNativeUtils.AddScaleU(DefaultScale, psrc, pdst, Length);
+                Thunk.AddScaleU(DefaultScale, psrc, pdst, Length);
             }
         }
 
@@ -65,7 +76,7 @@ namespace Microsoft.ML.CpuMath.PerformanceTests
             fixed (float* pdst = dst)
             fixed (int* pidx = idx)
             {
-                CpuMathNativeUtils.AddScaleSU(DefaultScale, psrc, pidx, pdst, IndexLength);
+                Thunk.AddScaleSU(DefaultScale, psrc, pidx, pdst, IndexLength);
             }
         }
 
@@ -76,7 +87,7 @@ namespace Microsoft.ML.CpuMath.PerformanceTests
             fixed (float* pdst = dst)
             fixed (float* pres = result)
             {
-                CpuMathNativeUtils.AddScaleCopyU(DefaultScale, psrc, pdst, pres, Length);
+                Thunk.AddScaleCopyU(DefaultScale, psrc, pdst, pres, Length);
             }
         }
 
@@ -86,7 +97,7 @@ namespace Microsoft.ML.CpuMath.PerformanceTests
             fixed (float* psrc = src)
             fixed (float* pdst = dst)
             {
-                CpuMathNativeUtils.AddU(psrc, pdst, Length);
+                Thunk.AddU(psrc, pdst, Length);
             }
         }
 
@@ -97,7 +108,7 @@ namespace Microsoft.ML.CpuMath.PerformanceTests
             fixed (float* pdst = dst)
             fixed (int* pidx = idx)
             {
-                CpuMathNativeUtils.AddSU(psrc, pidx, pdst, IndexLength);
+                Thunk.AddSU(psrc, pidx, pdst, IndexLength);
             }
         }
 
@@ -108,7 +119,7 @@ namespace Microsoft.ML.CpuMath.PerformanceTests
             fixed (float* psrc2 = src2)
             fixed (float* pdst = dst)
             {
-                CpuMathNativeUtils.MulElementWiseU(psrc1, psrc2, pdst, Length);
+                Thunk.MulElementWiseU(psrc1, psrc2, pdst, Length);
             }
         }
 
@@ -117,7 +128,7 @@ namespace Microsoft.ML.CpuMath.PerformanceTests
         {
             fixed (float* psrc = src)
             {
-                return CpuMathNativeUtils.Sum(psrc, Length);
+                return Thunk.Sum(psrc, Length);
             }
         }
 
@@ -126,7 +137,7 @@ namespace Microsoft.ML.CpuMath.PerformanceTests
         {
             fixed (float* psrc = src)
             {
-                return CpuMathNativeUtils.SumSqU(psrc, Length);
+                return Thunk.SumSqU(psrc, Length);
             }
         }
 
@@ -135,7 +146,7 @@ namespace Microsoft.ML.CpuMath.PerformanceTests
         {
             fixed (float* psrc = src)
             {
-                return CpuMathNativeUtils.SumSqDiffU(DefaultScale, psrc, Length);
+                return Thunk.SumSqDiffU(DefaultScale, psrc, Length);
             }
         }
 
@@ -144,7 +155,7 @@ namespace Microsoft.ML.CpuMath.PerformanceTests
         {
             fixed (float* psrc = src)
             {
-                return CpuMathNativeUtils.SumAbsU(psrc, Length);
+                return Thunk.SumAbsU(psrc, Length);
             }
         }
 
@@ -153,7 +164,7 @@ namespace Microsoft.ML.CpuMath.PerformanceTests
         {
             fixed (float* psrc = src)
             {
-                return CpuMathNativeUtils.SumAbsDiffU(DefaultScale, psrc, Length);
+                return Thunk.SumAbsDiffU(DefaultScale, psrc, Length);
             }
         }
 
@@ -162,7 +173,7 @@ namespace Microsoft.ML.CpuMath.PerformanceTests
         {
             fixed (float* psrc = src)
             {
-                return CpuMathNativeUtils.MaxAbsU(psrc, Length);
+                return Thunk.MaxAbsU(psrc, Length);
             }
         }
 
@@ -171,7 +182,7 @@ namespace Microsoft.ML.CpuMath.PerformanceTests
         {
             fixed (float* psrc = src)
             {
-                return CpuMathNativeUtils.MaxAbsDiffU(DefaultScale, psrc, Length);
+                return Thunk.MaxAbsDiffU(DefaultScale, psrc, Length);
             }
         }
 
@@ -181,7 +192,7 @@ namespace Microsoft.ML.CpuMath.PerformanceTests
             fixed (float* psrc = src)
             fixed (float* pdst = dst)
             {
-                return CpuMathNativeUtils.DotU(psrc, pdst, Length);
+                return Thunk.DotU(psrc, pdst, Length);
             }
         }
 
@@ -192,7 +203,7 @@ namespace Microsoft.ML.CpuMath.PerformanceTests
             fixed (float* pdst = dst)
             fixed (int* pidx = idx)
             {
-                return CpuMathNativeUtils.DotSU(psrc, pdst, pidx, IndexLength);
+                return Thunk.DotSU(psrc, pdst, pidx, IndexLength);
             }
         }
 
@@ -202,7 +213,7 @@ namespace Microsoft.ML.CpuMath.PerformanceTests
             fixed (float* psrc = src)
             fixed (float* pdst = dst)
             {
-                return CpuMathNativeUtils.Dist2(psrc, pdst, Length);
+                return Thunk.Dist2(psrc, pdst, Length);
             }
         }
 
@@ -213,7 +224,7 @@ namespace Microsoft.ML.CpuMath.PerformanceTests
             fixed (float* pdst = dst)
             fixed (float* pres = result)
             {
-                CpuMathNativeUtils.SdcaL1UpdateU(DefaultScale, psrc, DefaultScale, pdst, pres, Length);
+                Thunk.SdcaL1UpdateU(DefaultScale, psrc, DefaultScale, pdst, pres, Length);
             }
         }
 
@@ -225,20 +236,36 @@ namespace Microsoft.ML.CpuMath.PerformanceTests
             fixed (float* pres = result)
             fixed (int* pidx = idx)
             {
-                CpuMathNativeUtils.SdcaL1UpdateSU(DefaultScale, psrc, pidx, DefaultScale, pdst, pres, IndexLength);
+                Thunk.SdcaL1UpdateSU(DefaultScale, psrc, pidx, DefaultScale, pdst, pres, IndexLength);
             }
         }
 
         [Benchmark]
-        public void MatMul()
-            =>  SseUtils.MatTimesSrc(false, testMatrixAligned, testSrcVectorAligned, testDstVectorAligned, matrixLength);
-        
+        public unsafe void MatMul()
+        {
+            fixed (float* pmat = &testMatrixAligned.Items[0])
+            fixed (float* psrc = &testSrcVectorAligned.Items[0])
+            fixed (float* pdst = &testDstVectorAligned.Items[0])
+                Thunk.MatMul(Ptr(testMatrixAligned, pmat), Ptr(testSrcVectorAligned, psrc), Ptr(testDstVectorAligned, pdst), matrixLength, testSrcVectorAligned.Size);
+        }
+
         [Benchmark]
-        public void MatMulTran()
-            =>  SseUtils.MatTimesSrc(true, testMatrixAligned, testSrcVectorAligned, testDstVectorAligned, matrixLength);     
+        public unsafe void MatMulTran()
+        {
+            fixed (float* pmat = &testMatrixAligned.Items[0])
+            fixed (float* psrc = &testSrcVectorAligned.Items[0])
+            fixed (float* pdst = &testDstVectorAligned.Items[0])
+                Thunk.MatMulTran(Ptr(testMatrixAligned, pmat), Ptr(testSrcVectorAligned, psrc), Ptr(testDstVectorAligned, pdst), testDstVectorAligned.Size, matrixLength);
+        }
 
         [Benchmark]
         public unsafe void MatMulP()
-            =>  SseUtils.MatTimesSrc(testMatrixAligned, matrixIdx, testSrcVectorAligned, 0, 0, MatrixIndexLength, testDstVectorAligned, matrixLength);     
+        {
+            fixed (float* pmat = &testMatrixAligned.Items[0])
+            fixed (float* psrc = &testSrcVectorAligned.Items[0])
+            fixed (float* pdst = &testDstVectorAligned.Items[0])
+            fixed (int* ppossrc = &matrixIdx[0])
+                Thunk.MatMulP(Ptr(testMatrixAligned, pmat), ppossrc, Ptr(testSrcVectorAligned, psrc), 0, 0, MatrixIndexLength, Ptr(testDstVectorAligned, pdst), matrixLength, testSrcVectorAligned.Size);
+        }
     }
 }
